requirements.txt
python-dotenv
Tweepy
Twarc
